scripts: diff COUNTER-Robots and Make-Data-Count lists
Make-Data-Count uses a non up-to-date version of the COUNTER-Robots
list. However, the differentiation between robots and machines is nice.
To make sure that we have the latest list, but with differentiation we
diff the COUNTER-Robots robots.txt with the machines.txt and
unclassified.txt from the Make-Data-Count repository. If the resulting
classification contains errors, than this must be fixed by opening a PR
at either source repository.

Testruns on Python 2.7 had the following error during pip isntall:

    ERROR: pytest-cov 2.7.1 has requirement pytest>=3.6,
    but you'll have pytest 3.3.2 which is incompatible.

This caused problems during the test. Bumping the pytest version
resolves this.