global: automatic deploy of cron triggered builds
For the automatic deploy of cron triggered builds we added two scripts
and changed the layout of the travis jobs. The scripts we added:

run-update.sh triggers the update script for the user-agant lists. If an
update is detect the changeg user-agent lists are commited. We also
update the version of the package and create a seperate, tagged, commit
for the release.

push-update.sh adds the remote branch to the current environment
(required for Travis builds) and pushes all changes including tags to
the remote.

For the Travis builds we decided to destinguish three different
situations:
1. Regular build
2. Cron triggered build
3. Tagged build

Regular builds run the test on three different Python versions: 2.7, 3.5
and 3.6

Cron triggered jobs only check if there are updates and do a simple test
before pushing.

Tagged builds first run the Test stage, similar to a regular build.
After succes an addtional stage is triggered, that tests and creates the
release.